클래스 리스트 프론트, 클래스 상세페이지 적용

+++ src/main/java/kr/or/eum/product/model/vo/Product.java
@@ -23,7 +23,7 @@ public class Product {
 	private String productAns;
 	private String productAddr;
 	private int sale;
-	private int reviewStar;
+	private double reviewStar;
 	private int revCount;
 	private int productOption;
